New: Type hints on record.py
This MR adds type hints to record.py. It currently passes with my local mypy installation.

I have attempted to not change any code behaviours, but this has meant that I have had to tell mypy to ignore some lines. For example, `self["245"]["a"]` does not pass the type checker, since `self["245"]` could be `None`, and that is not indexable. In other cases, particularly around the typing of `Leader`, I have used the `Any` type. (mypy seems to get confused over whether it's a string or a Leader object, and cannot automatically infer this).

(If someone else wanted to have a look at these behaviours and make them more typeable, then they should be able to find them quite easily by looking for `# type: ignore` in the code)

I also noticed an error in the typing for `__getitem__` in the Leader -- this can take an int, slice, or string, and not just a string. This has been fixed as well.