Fixed: Iterating over fields fails
Previously, iterating over fields and calling 'field.code' on the list of fields might fail, if the field was a control field.

This was because `self.subfields` was only initialized if the field was not a control field, so it was attempting to iterate over an undefined property. For any blind iteration of all fields in a record, and all subfields in those fields, this raised an exception.

To fix this, all fields are now pre-initialized in the init method, and updated with their value depending on the field status as a control field. Iterating over the subfields of a control field will no longer crash, but the list of subfields will be empty. Similarly, the `data` property is initialized to `None`, so accessing non-control fields' data property will also not crash.

Additionally, the `is_control_field()` method has been refactored, and now takes its value from the `control_field` property set on initialization. This is because the nature of the field is set at init, so no need to re-check it.