MARCMakerReader
This commit adds a new class pymarc.str.MARCMakerReader for reading MARC records
that are encoded using the MARCMaker format. Previously pymarc only had
the ability to *write* records in this format.

MARCMaker is human readable version of a MARC record, for example:

```text
=LDR  01060cam  22002894a 4500
=001  11778504
=008  990802s2000\\\\mau\\\\\\b\\\\001\0\eng\\
=050  00$aQA76.6$b.H857 2000
=100  1\$aHunt, Andrew,$d1964-
=245  14$aThe pragmatic programmer :$bfrom journeyman to master /$cAndrew Hunt, David Thomas.
=260  \\$aReading, Mass :$bAddison-Wesley,$c2000.
=300  \\$axxiv, 321 p. ;$c24 cm.
=504  \\$aIncludes bibliographical references.
=650  \0$aComputer programming.
=700  1\$aThomas, David,$d1956-
```

In order to read it you instantiate a MARCMakerReader object and use it
as an iterator. For example, to get the first record:

```python
from pymarc import MARCMakerReader

rec = next(MARCMakerReader(open('record.mkr')))
print(rec.title)
```

It works similarl to MARCReader, but you don't open the file in binary
mode, since it's assume to be a UTF-8 encoded string. You can change the
encoding with the `encoding` parameter.