Test the Record fields parameter
I might be mistaken, but I didn't see a test for the new fields
parameter on Record.__init__, so this commit adds one.

In addition I noticed that in CI the call to ruff check wasn't returning
zero for some errors. So I fixed that and cleaned up some more lint
issues in the tests.